Create react app for Week 9 HW
"This assignment calls for the table rendering of user data retrieved and parsed from a remote API endpoint. Your job is to understand the following:

How Promise works with data fetching remotely.
How Table components work in relation to other UI components that are children of Table.
Then, once you understand the above, replace YOUR_CODE_HERE with the correct code to render the UI as shown below."